+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Version 0.2.6
 - Add `Engine.getServices` to retrieve service subclasses
+- Made `FamilyPresenceWatcherSystem` more open for extensions
 
 # Version 0.2.5 (released 2020-01-21)
 - `EngineService`s are now automatically wired and initialized